encoders/image/FaceNetEncoder/requirements.txt
facenet-pytorch==2.5.2
torchvision==0.9.0
